Fix module icon and number showing up at same time
If the user tapped a module while the module info number was doing its
"zoom in" animation, both the info number and the icon would show up at
the same time. This was caused by the ng-hide on the info number waiting
until the info number animation finished to actually hide the element.
The module directive now has a timer that blocks interaction on the
module while the info number is animating.