Create Gradle properties file if it does not exist
The Gradle properties file forces the creation of two separate APKs. It
must exist in the platforms/android directory, which is removed entirely
when removing/adding the Android platform (common troubleshooting step).
The script will now create the gradle.properties file each time it runs
if the file does not already exist.