Refactor loginController to use DbFactory
Previously, the loginController would store credentials using
StorageService before checking if they were valid or not. This was
intended to take advantage of the Base64 encoding of the password that
occurs in StorageService.storeCredentials(). I abstracted password
encoding to Base64 into a $filter, which is now used in the login()
function in this controller. This has the added benefit of reducing the
amount of time that the user's password is stored in plaintext.