Refactor banner refresh function into directive without database usage
Previously, the banner would rely on DataService to tell it whether or
not enough time had passed to merit an AppInfo refresh. The time of the
last app info refresh was cached in localStorage. Now, the time of the
last refresh is saved in a variable because it does not have to persist
between sessions. The directive now directly uses DataService.get().