Remove unneccesary logic from appBanner
The directive was doing extra checks to make sure that the app
info actually needed to be refreshed before it actually made the
request. This is a holdover from when the app info was refreshed
whenever the user tapped a module to load it. If the user tapped a bunch
in a row, there would be many unnecessary requests for app info in a row.
However, the logic now simply refreshes app info on load and on an
interval defined by AppInfoRefreshTime after that.