Fix Highland Express schedule scrolling
The schedule added several destinations but was still constrained to
the width of the screen, so the new destinations were cut off. This
commit adds support for horizontal scrolling in the schedule view. The
<ion-scroll> directive did not work as a drop-in like it should have,
hence the getDimensions() function and the inline styles in the view.
This solution automatically sizes the schedule <ion-scroll> based on
the number of rows and columns in the schedule.