Merge dependencies and bump all to latest version
Since npm is only used for local development purposes in this project,
all dependencies can be considered dev dependencies. Being dev
dependencies, it makes since for them to always be at the latest
version, as shown by the number of errors from outdated packages that
came up after not running the project for months.