Fix user privacy policy requests
Setting the privacy policy was failing with a `409 Conflict` error code
because the user data sent back to the API was stale. This problem could
be fixed with better passing of data from `LoginController` to
`PrivacyPolicyController`, but is good enough as is because the user
will only have to accept the privacy policy once.